refactor(core): split the three files that outgrew a thousand lines
paginated-surface.ts keeps the composition root and gains siblings for
its host contract, input wiring, selection/op planning, formatting
queries and page environment; ooxml-tree.ts keeps types and the read
path with serializer, validator and shared QName helpers extracted
(leaf ooxml-shared.ts keeps the import graph acyclic); tree-ops.ts
splits op validation from op application behind an unchanged entry.
Behavior-preserving: barrels and importers untouched, no new public API.
