+++ packages/core/src/store/store/tree-op-validate.ts
@@ -1,0 +1,330 @@
+// The op vocabulary and pre-application validation (tree-ops seam).
+//
+// This module owns what an op IS — the declarative, JSON-safe `TreeDocOp` shapes, the
+// accepted property boundaries, the effect/rejection contracts — plus the segment model
+// that flattens a paragraph into UTF-16 addressable units, and `validateTreeOp`, which
+// runs BEFORE any tree work so a rejected op leaves the tree, revision and indexes exactly
+// as they were. Application lives in tree-op-apply.ts; both are re-exported via tree-ops.ts.
+
+import type { OoxmlNode, OoxmlParagraphNode, OoxmlPart } from '../package/ooxml-tree.ts';
+import { findNode } from '../package/ooxml-edit.ts';
+import { isValidXmlText } from '../package/sinks.ts';
+
+/**
+ * The accepted RUN property boundary (design D8), as the OOXML element names that carry it.
+ *
+ * An explicit allowlist rather than "any `w:rPr` child": a property outside D8 has no
+ * resolver, no layout behavior and no support claim, so accepting it here would let an
+ * operation assert support the engine does not have. Unknown properties still ROUND-TRIP —
+ * they are generic nodes in the tree — they simply cannot be authored by an op.
+ */
+export const ACCEPTED_RUN_PROPERTIES = [
+  'rFonts', // font family
+  'sz', // half-point size
+  'szCs',
+  'color',
+  'b', // bold
+  'bCs',
+  'i', // italic
+  'iCs',
+  'u', // underline variant and color
+  'strike',
+  'dstrike', // double strike
+  'highlight',
+  'vertAlign', // superscript / subscript
+  'position', // baseline offset
+  'caps',
+  'smallCaps',
+  'spacing', // character spacing
+  'w', // horizontal scaling
+  'kern',
+] as const;
+
+/** The accepted PARAGRAPH property boundary (design D8). */
+export const ACCEPTED_PARAGRAPH_PROPERTIES = [
+  'pStyle',
+  'jc', // alignment
+  'spacing', // before/after + line spacing and rule
+  'ind', // left/right/first-line/hanging indents
+  'tabs',
+  'numPr', // numbering identity and level
+  'keepNext',
+  'keepLines',
+  'widowControl',
+  'pageBreakBefore',
+  'shd', // shading
+] as const;
+
+export type AcceptedRunProperty = (typeof ACCEPTED_RUN_PROPERTIES)[number];
+export type AcceptedParagraphProperty = (typeof ACCEPTED_PARAGRAPH_PROPERTIES)[number];
+
+const RUN_PROPERTY_SET: ReadonlySet<string> = new Set(ACCEPTED_RUN_PROPERTIES);
+const PARAGRAPH_PROPERTY_SET: ReadonlySet<string> = new Set(ACCEPTED_PARAGRAPH_PROPERTIES);
+
+/**
+ * One authored property: an element name plus its `w:`-namespace attributes.
+ *
+ * Modeled as name+attributes rather than a typed record per property because that is what
+ * the tree holds, so an op maps to nodes without a lossy intermediate vocabulary. Attribute
+ * VALUES are validated as XML text; their meaning is the resolver's business.
+ */
+export interface OoxmlProperty {
+  readonly localName: string;
+  readonly attributes?: Readonly<Record<string, string>>;
+}
+
+export type TreeDocOp =
+  | {
+      readonly op: 'insertText';
+      readonly paragraphId: string;
+      readonly offset: number;
+      readonly text: string;
+    }
+  | {
+      readonly op: 'deleteText';
+      readonly paragraphId: string;
+      readonly start: number;
+      readonly end: number;
+    }
+  | { readonly op: 'insertTab'; readonly paragraphId: string; readonly offset: number }
+  | { readonly op: 'insertHardBreak'; readonly paragraphId: string; readonly offset: number }
+  | { readonly op: 'splitParagraph'; readonly paragraphId: string; readonly offset: number }
+  | {
+      /**
+       * Split one `w:p` at MANY offsets in a single op.
+       *
+       * Equivalent to applying `splitParagraph` at each offset from the last to the first,
+       * but the paragraph's content is cut in one pass and the parent's child sequence is
+       * rebuilt once. A plain-text paste is a paragraph mark per line: as individual ops,
+       * a large paste rebuilt the body — and re-sliced the pasted text — once per line,
+       * which is quadratic in paste size.
+       */
+      readonly op: 'splitParagraphMany';
+      readonly paragraphId: string;
+      /**
+       * Non-decreasing UTF-16 offsets; each produces one paragraph boundary. A repeated
+       * offset produces an empty paragraph between the two boundaries — a blank line.
+       */
+      readonly offsets: readonly number[];
+    }
+  | { readonly op: 'joinParagraphs'; readonly firstId: string; readonly secondId: string }
+  | {
+      readonly op: 'setRunProperties';
+      readonly paragraphId: string;
+      readonly start: number;
+      readonly end: number;
+      readonly properties: readonly OoxmlProperty[];
+    }
+  | {
+      readonly op: 'setParagraphProperties';
+      readonly paragraphId: string;
+      readonly properties: readonly OoxmlProperty[];
+    };
+
+export type TreeDocOpKind = TreeDocOp['op'];
+
+export const TREE_DOC_OP_KINDS = [
+  'insertText',
+  'deleteText',
+  'insertTab',
+  'insertHardBreak',
+  'splitParagraph',
+  'splitParagraphMany',
+  'joinParagraphs',
+  'setRunProperties',
+  'setParagraphProperties',
+] as const satisfies readonly TreeDocOpKind[];
+
+// Compile-time exhaustiveness, matching the legacy `DOC_OP_KINDS` guard: a new op must be
+// listed here or this fails to typecheck, so it can never be silently unvalidated.
+type _MissingTreeOp = Exclude<TreeDocOpKind, (typeof TREE_DOC_OP_KINDS)[number]>;
+const _treeOpsExhaustive: _MissingTreeOp extends never ? true : ['missing', _MissingTreeOp] = true;
+void _treeOpsExhaustive;
+
+/**
+ * How far a committed op can reach, so layout can scope its work (task 5.2).
+ *
+ * `text-local` touches one paragraph's characters; `paragraph-local` changes one
+ * paragraph's own properties; `flow-structural` changes the block sequence and can
+ * repaginate everything after it.
+ */
+export type ImpactClass = 'text-local' | 'paragraph-local' | 'flow-structural';
+
+export interface TreeOpEffect {
+  readonly dirty: readonly string[];
+  readonly created: readonly string[];
+  readonly deleted: readonly string[];
+  readonly split?: { readonly from: string; readonly tail: string };
+  /** One entry per boundary of a many-way split, in document order. */
+  readonly splits?: readonly { readonly from: string; readonly tail: string }[];
+  readonly join?: { readonly kept: string; readonly removed: string };
+  readonly dependencyKeys: readonly string[];
+  readonly impact: ImpactClass;
+}
+
+export type TreeOpRejection =
+  | 'unknown-op'
+  | 'unknown-paragraph'
+  | 'not-a-paragraph'
+  | 'offset-out-of-range'
+  | 'invalid-range'
+  | 'splits-surrogate-pair'
+  | 'invalid-text'
+  | 'unsupported-property'
+  | 'invalid-property-value'
+  | 'not-adjacent-siblings'
+  | 'tree-invariant';
+
+export type TreeOpResult =
+  | { readonly ok: true; readonly part: OoxmlPart; readonly effect: TreeOpEffect }
+  | { readonly ok: false; readonly reason: TreeOpRejection; readonly detail?: string };
+
+/** One addressable unit of paragraph text: a text value, a tab, or a hard break. */
+export interface Segment {
+  readonly runId: string;
+  readonly node: OoxmlNode;
+  readonly start: number;
+  readonly end: number;
+}
+
+export function isParagraph(node: OoxmlNode | null): node is OoxmlParagraphNode {
+  return node !== null && node.kind === 'paragraph';
+}
+
+/** Flatten a paragraph into UTF-16 addressable segments, in document order. */
+export function segmentsOf(paragraph: OoxmlParagraphNode): Segment[] {
+  const segments: Segment[] = [];
+  let offset = 0;
+  const visit = (node: OoxmlNode, runId: string): void => {
+    if (node.kind === 'textValue') {
+      segments.push({ runId, node, start: offset, end: offset + node.value.length });
+      offset += node.value.length;
+      return;
+    }
+    if (node.kind === 'tab' || node.kind === 'hardBreak') {
+      segments.push({ runId, node, start: offset, end: offset + 1 });
+      offset += 1;
+      return;
+    }
+    if (node.kind === 'runProperties' || node.kind === 'generic') return;
+    for (const child of node.children) visit(child, runId);
+  };
+  for (const child of paragraph.children) {
+    if (child.kind !== 'run') continue;
+    for (const grand of child.children) visit(grand, child.id);
+  }
+  return segments;
+}
+
+function paragraphLength(paragraph: OoxmlParagraphNode): number {
+  const segments = segmentsOf(paragraph);
+  return segments.length === 0 ? 0 : segments[segments.length - 1]!.end;
+}
+
+/** Whether an offset falls between the halves of a surrogate pair. */
+function splitsSurrogate(paragraph: OoxmlParagraphNode, offset: number): boolean {
+  for (const segment of segmentsOf(paragraph)) {
+    if (segment.node.kind !== 'textValue') continue;
+    if (offset <= segment.start || offset >= segment.end) continue;
+    const local = offset - segment.start;
+    const before = segment.node.value.charCodeAt(local - 1);
+    const after = segment.node.value.charCodeAt(local);
+    if (before >= 0xd800 && before <= 0xdbff && after >= 0xdc00 && after <= 0xdfff) return true;
+  }
+  return false;
+}
+
+function validateProperties(
+  properties: readonly OoxmlProperty[],
+  accepted: ReadonlySet<string>
+): TreeOpRejection | null {
+  for (const property of properties) {
+    if (!accepted.has(property.localName)) return 'unsupported-property';
+    for (const [name, value] of Object.entries(property.attributes ?? {})) {
+      // Attribute names and values are written straight into XML on save, so both are
+      // checked here rather than at the sink — a rejected op must never reach the tree.
+      if (!/^[A-Za-z_][\w.-]*$/.test(name)) return 'invalid-property-value';
+      if (typeof value !== 'string' || !isValidXmlText(value)) return 'invalid-property-value';
+    }
+  }
+  return null;
+}
+
+/** Structural validation, run before any tree work so a rejection changes nothing. */
+export function validateTreeOp(part: OoxmlPart, op: TreeDocOp): TreeOpRejection | null {
+  if (!TREE_DOC_OP_KINDS.includes(op.op)) return 'unknown-op';
+
+  if (op.op === 'joinParagraphs') {
+    const first = findNode(part, op.firstId);
+    const second = findNode(part, op.secondId);
+    if (!first || !second) return 'unknown-paragraph';
+    if (!isParagraph(first) || !isParagraph(second)) return 'not-a-paragraph';
+    return null;
+  }
+
+  const paragraph = findNode(part, op.paragraphId);
+  if (!paragraph) return 'unknown-paragraph';
+  if (!isParagraph(paragraph)) return 'not-a-paragraph';
+  const length = paragraphLength(paragraph);
+
+  switch (op.op) {
+    case 'insertText': {
+      if (!Number.isInteger(op.offset) || op.offset < 0 || op.offset > length) {
+        return 'offset-out-of-range';
+      }
+      if (typeof op.text !== 'string' || !isValidXmlText(op.text)) return 'invalid-text';
+      if (splitsSurrogate(paragraph, op.offset)) return 'splits-surrogate-pair';
+      return null;
+    }
+    case 'insertTab':
+    case 'insertHardBreak': {
+      if (!Number.isInteger(op.offset) || op.offset < 0 || op.offset > length) {
+        return 'offset-out-of-range';
+      }
+      if (splitsSurrogate(paragraph, op.offset)) return 'splits-surrogate-pair';
+      return null;
+    }
+    case 'splitParagraph': {
+      if (!Number.isInteger(op.offset) || op.offset < 0 || op.offset > length) {
+        return 'offset-out-of-range';
+      }
+      if (splitsSurrogate(paragraph, op.offset)) return 'splits-surrogate-pair';
+      return null;
+    }
+    case 'splitParagraphMany': {
+      if (!Array.isArray(op.offsets) || op.offsets.length === 0) return 'invalid-range';
+      let previous = -1;
+      for (const offset of op.offsets) {
+        if (!Number.isInteger(offset) || offset < 0 || offset > length) {
+          return 'offset-out-of-range';
+        }
+        // Non-decreasing: unordered offsets have no single sequential reading, but a
+        // REPEATED offset does — it is how an empty paragraph is expressed, and a paste
+        // with a blank line carries exactly that.
+        if (offset < previous) return 'invalid-range';
+        previous = offset;
+        if (splitsSurrogate(paragraph, offset)) return 'splits-surrogate-pair';
+      }
+      return null;
+    }
+    case 'deleteText': {
+      if (!Number.isInteger(op.start) || !Number.isInteger(op.end)) return 'invalid-range';
+      if (op.start < 0 || op.end > length) return 'offset-out-of-range';
+      if (op.start >= op.end) return 'invalid-range';
+      if (splitsSurrogate(paragraph, op.start) || splitsSurrogate(paragraph, op.end)) {
+        return 'splits-surrogate-pair';
+      }
+      return null;
+    }
+    case 'setRunProperties': {
+      if (!Number.isInteger(op.start) || !Number.isInteger(op.end)) return 'invalid-range';
+      if (op.start < 0 || op.end > length) return 'offset-out-of-range';
+      if (op.start >= op.end) return 'invalid-range';
+      return validateProperties(op.properties, RUN_PROPERTY_SET);
+    }
+    case 'setParagraphProperties':
+      return validateProperties(op.properties, PARAGRAPH_PROPERTY_SET);
+    default:
+      return 'unknown-op';
+  }
+}
